requirements.txt
# 工具检测系统 - 核心依赖包

# 核心机器学习库 (必需)
torch>=1.13.0
open-clip-torch>=2.0.0
numpy<2.0.0
pillow>=9.0.0

# 视频流处理 (视频检测功能必需)
opencv-python>=4.5.0

# Web服务 (可选 - 启用Web API时安装)
# flask>=2.0.0
# flask-cors>=3.0.0
# requests>=2.28.0

# 实时监控 (可选 - 启用监控系统时安装)  
# schedule>=1.2.0

# 命令行工具美化 (可选 - 启用彩色CLI时安装)
# tabulate>=0.9.0
# colorama>=0.4.0